+++ sdk/resourcemanager/databox/armdatabox/zz_generated_management_client.go
@@ -96,7 +96,7 @@ func (client *ManagementClient) mitigateCreateRequest(ctx context.Context, jobNa
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header.Set("Accept", "application/json")
 	return req, runtime.MarshalAsJSON(req, mitigateJobRequest)